pyproject.toml
[project]
name = "bua"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "browserbase>=1.2.0",
    "halo>=0.0.31",
    "httpx>=0.28.1",
    "notte>=1.4.1",
    "pillow>=11.2.1",
    "playwright>=1.50.0",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "scrapybara>=2.5.1",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
